Reorganize sampler tests into per-stage package, prune redundancy, add gap tests
Split the 1.8k-line tests/test_sample_tasks.py monolith into a per-stage
tests/sampler/ package that maps 1:1 onto the 5-stage pipeline, prune
low-value/redundant tests, and add targeted gap coverage.

Layout (tests/sampler/):
- test_stage0_prediction_times.py  build_prediction_times (+ invariants)
- test_stage1_query_distribution.py  QueryDistribution
- test_stage2_patient_contexts.py  sample_patient_contexts
- test_stage3_build_index.py  build_index
- test_stage4_labeling.py  evaluate_index_df + label_one_shard
- test_pure_helpers.py  primitives, path/config resolution, atomic writes, workers
- test_orchestration.py  run() end-to-end, cross-process determinism, snapshot
- conftest.py  shared fixtures/helpers

Pruned (~9 tests, no coverage loss): trivial path-string equalities already
covered by source doctests; redundant resolve-path wrapper cases; the
tempfile.mkstemp uniqueness test; brittle config-literal change-detectors
(legacy-keys-gone, defaults-match-spec) replaced by a presence+type check.

Added gap tests: log-uniform skews shorter + code-draw coverage (stage 1);
subject-universe coverage + nonempty-path index dtype (stage 2); artifact-level
determinism + cross-shard zip alignment (stage 3); forward-only asof (past
events ignored), query-code isolation, multiple-matches-still-true (stage 4);
the row-count guard actually raises + overwrite forces rebuild (orchestration).
Also strengthened the atomic-write test to assert the failure path leaves no
orphan temp.

The sampler unit layer no longer depends on the HF-backed demo_model fixture:
a local conftest overrides the repo-root autouse doctest-namespace fixture so
these pure data-transform tests run fast and offline. The one test that needs
the model (downstream dataset drop-in) moves to top-level
tests/test_sampler_dataset_integration.py, where the root fixtures still apply.

111 sampler tests pass offline.
